requirements.txt
https://github.com/LonamiWebs/Telethon/archive/master.zip
telethon >= 1.16.1
aiofiles
aiohttp
aria2p
asyncurban
beautifulsoup4
cairosvg
cfscrape
coffeehouse
covid
cowpy
cryptg
csvfaker
DateTime
fontTools
pymongo
python-dotenv
emoji
fonttools
geopy
gitpython
glitch_this
google-api-python-client==1.7.11
google_images_download
googletrans
gsearch
gtts
hachoir
heroku3
httplib2<0.16.0
humanize
jotquote
justwatch
kwot
lottie
lxml
lyricsgenius
nekos.py
oauth2client
patool
Pillow
programmingquotes
psutil
psycopg2
pybase64
pyDownload
pyfiglet
PyGithub
pylast
PyLyrics
pySmartDL
python-barcode
python-magic
pytube>=9.5.1
pytuneteller
pytz
qrcode
regex
requests
search-engine-parser
selenium
speedtest-cli
spotify_token
sqlalchemy
telegraph
tswift
urbandict
validators
wand
wikipedia
youtube-dl
